# Conflicts:
#	.gitignore

+++ egg/app/controller/member.js
@@ -1,8 +1,8 @@
-'use strict';
+'use strict'
 
-const baseController = require('./baseController');
-const Sequelize = require('sequelize');
-const Op = Sequelize.Op;
+const baseController = require('./baseController')
+const Sequelize = require('sequelize')
+const Op = Sequelize.Op
 
 class MemberController extends baseController {
 
@@ -11,12 +11,40 @@ class MemberController extends baseController {
      * @returns {Promise<void>}
      */
     async create() {
-        const {ctx} = this;
-        const result = await ctx.model.Member.create({
-            ...ctx.request.body,
-            password: ctx.setToken({password: ctx.randomString()})
-        })
-        this.result({data: result})
+        const {ctx} = this
+        const body = ctx.request.body
+        try {
+            this.ctx.validate({
+                username: {type: 'string', min: 2, max: 20, require: true},
+            })
+
+            const resultUsername = await ctx.model.Member.findOne({
+                where: {
+                    username: body.username,
+                },
+            })
+            if (resultUsername) {
+                this.result({data: '', message: '该用户名已存在', code: 1003})
+            } else {
+                let password = body.password
+                if (password) {
+                    this.ctx.validate({
+                        password: {type: 'string', min: 2, max: 20, require: true},
+                    })
+                } else {
+                    password = ctx.randomString()
+                }
+                body.roles = 15; // 默认角色
+                const result = await ctx.model.Member.create({
+                    ...ctx.request.body,
+                    password: ctx.setToken({password: password, username: ctx.request.body.username}),
+                })
+                this.result({data: result})
+            }
+        } catch (error) {
+            const {errors = []} = error
+            this.result({data: '', message: errors[0].field + ' ' + errors[0].message, code: 1001})
+        }
     }
 
     /**
@@ -24,12 +52,16 @@ class MemberController extends baseController {
      * @returns {Promise<void>}
      */
     async delete() {
-        const {ctx} = this;
-        const {id} = ctx.request.body;
-        const result = await ctx.model.Member.destroy({
-            where: {id}
-        })
-        this.result({data: result})
+        const {ctx} = this
+        const {id} = ctx.request.body
+        if (Number(id) === 1) {
+            this.result({data: '', message: '禁止删除该条数据', code: 1001})
+        } else {
+            const result = await ctx.model.Member.destroy({
+                where: {id},
+            })
+            this.result({data: result})
+        }
     }
 
     /**
@@ -37,12 +69,12 @@ class MemberController extends baseController {
      * @returns {Promise<void>}
      */
     async deletes() {
-        const {ctx} = this;
-        const {ids} = ctx.request.body;
+        const {ctx} = this
+        const {ids} = ctx.request.body
         const result = await ctx.model.Member.destroy({
             where: {
-                id: [...ids]
-            }
+                id: [...ids],
+            },
         })
         this.result({data: result})
     }
@@ -53,11 +85,15 @@ class MemberController extends baseController {
      * @returns {Promise<void>}
      */
     async find() {
-        const {ctx} = this;
-        const {id} = ctx.request.body;
+        const {ctx} = this
+        const {id} = ctx.request.body
         const result = await ctx.model.Member.findOne({
-            where: {id}
+            where: {id},
         })
+        const rolesDetail = await ctx.model.Role.findOne({
+            where: {id: result.roles},
+        })
+        result.setDataValue('rolesDetail', rolesDetail)
         this.result({data: result})
     }
 
@@ -66,34 +102,26 @@ class MemberController extends baseController {
      * @returns {Promise<void>}
      */
     async all() {
-        const {ctx} = this;
-        const {ks} = ctx.request.body;
+        const {ctx} = this
+        const param = ctx.request.body || {}
         const where = {}
-        if (ks) {
-            where.name = {[Op.like]: `%${ks}%`} // 模糊查詢 https://www.sequelize.com.cn/core-concepts/model-querying-basics
+        for (const paramKey in param) {
+            where[paramKey] = { [Op.like]: `%${param[paramKey]}%` } // 模糊查找
         }
         // attributes : 查询字段信息，可以通过 exclude 来指定不查询的字段，或者可以直接传入数组查询数组中的字段( attributes: ['id', 'nickname'] )
         // include : 指定查询的关联，as 必须和定义关联关系的 as 相同，model 指定关联的模型
         const result = await ctx.model.Member.findAll({
             where: {...where},
-            include: [{model: ctx.model.Branch, as: 'branch'}]
+            include: [{model: ctx.model.Branch, as: 'branch'}],
         })
 
         for (let i = 0; i < result.length; i++) {
-            const item = result[i];
-            if (item.roles) {
-                const arr = item.roles.split(',');
-                const roleResult = [];
-                for (let j = 0; j < arr.length; j++) {
-                    const obj = await ctx.model.Role.findOne({
-                        where: {tag: arr[j]}
-                    })
-                    roleResult.push(obj)
-                }
-                item.roles = roleResult;
-            }
+            const item = result[i]
+            const rolesDetail = await ctx.model.Role.findOne({
+                where: {id: item.roles},
+            })
+            item.setDataValue('rolesDetail', rolesDetail)
         }
-
         this.result({data: result})
     }
 
@@ -103,14 +131,14 @@ class MemberController extends baseController {
      * @returns {Promise<void>}
      */
     async update() {
-        const {ctx} = this;
-        const body = ctx.request.body;
+        const {ctx} = this
+        const body = ctx.request.body
         const result = await ctx.model.Member.update({
             ...body,
         }, {
             where: {
-                id: body.id
-            }
+                id: body.id,
+            },
         })
         this.result({data: result})
     }
@@ -120,39 +148,39 @@ class MemberController extends baseController {
      * @returns {Promise<void>}
      */
     async updatePas() {
-        const {ctx} = this;
-        const body = ctx.request.body;
+        const {ctx} = this
+        const body = ctx.request.body
         try {
             this.ctx.validate({
-                password: {type: 'string', min: 3, max: 20, require: true}
-            });
+                password: {type: 'string', min: 3, max: 20, require: true},
+            })
             const result = await ctx.model.Member.update({
-                password: ctx.setToken({password: body.password}),
+                password: ctx.setToken({password: body.password, username: body.username}),
             }, {
                 where: {
-                    id: body.id
-                }
+                    id: body.id,
+                },
             })
             this.result({data: result})
         } catch (error) {
-            const {errors = []} = error;
-            this.result({data: '', message: errors[0], code: 1001})
+            const {errors = []} = error
+            this.result({data: '', message: errors[0].field + ' ' + errors[0].message, code: 1001})
         }
     }
 
     async getPas() {
-        const {ctx} = this;
-        const body = ctx.request.body;
+        const {ctx} = this
+        const body = ctx.request.body
         try {
             this.ctx.validate({
-                password: {type: 'string', require: true}
-            });
+                password: {type: 'string', require: true},
+            })
             this.result({data: ctx.getDecodeToken(body.password)})
         } catch (error) {
-            const {errors = []} = error;
-            this.result({data: '', message: errors[0], code: 1001})
+            const {errors = []} = error
+            this.result({data: '', message: errors[0].field + ' ' + errors[0].message, code: 1001})
         }
     }
 }
 
-module.exports = MemberController;
+module.exports = MemberController
